+++ packages/at_contacts_flutter/lib/screens/contacts_screen.dart
@@ -1,7 +1,5 @@
-import 'package:at_contact/at_contact.dart';
-
 import 'package:at_common_flutter/at_common_flutter.dart';
-
+import 'package:at_contact/at_contact.dart';
 import 'package:at_contacts_flutter/models/contact_base_model.dart';
 import 'package:at_contacts_flutter/services/contact_service.dart';
 import 'package:at_contacts_flutter/utils/colors.dart';
@@ -34,9 +32,7 @@ class ContactsScreen extends StatefulWidget {
   const ContactsScreen(
       {Key? key,
       this.selectedList,
-      @Deprecated(
-          'context is no longer required and will be removed in upcoming version')
-      this.context,
+      @Deprecated('context is no longer required and will be removed in upcoming version') this.context,
       this.asSelectionScreen = false,
       this.asSingleSelectionScreen = false,
       this.saveGroup,
@@ -80,8 +76,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
 
       if (widget.selectedContactsHistory != null) {
         _contactService.selectedContacts = widget.selectedContactsHistory!;
-        _contactService.selectedContactSink
-            .add(_contactService.selectedContacts);
+        _contactService.selectedContactSink.add(_contactService.selectedContacts);
       }
     });
 
@@ -137,10 +132,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
           );
         },
         // ignore: unnecessary_null_comparison
-        showTrailingIcon: widget.asSelectionScreen == null ||
-                widget.asSelectionScreen == false
-            ? true
-            : false,
+        showTrailingIcon: widget.asSelectionScreen == null || widget.asSelectionScreen == false ? true : false,
         trailingIcon: const Center(
           child: Icon(
             Icons.add,
@@ -151,8 +143,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
       body: errorOcurred
           ? const ErrorScreen()
           : Container(
-              padding: EdgeInsets.symmetric(
-                  horizontal: 16.toWidth, vertical: 16.toHeight),
+              padding: EdgeInsets.symmetric(horizontal: 16.toWidth, vertical: 16.toHeight),
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.start,
                 children: [
@@ -175,8 +166,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
                     stream: _contactService.contactStream,
                     initialData: _contactService.baseContactList,
                     builder: (context, snapshot) {
-                      if ((snapshot.connectionState ==
-                          ConnectionState.waiting)) {
+                      if ((snapshot.connectionState == ConnectionState.waiting)) {
                         return const Center(
                           child: CircularProgressIndicator(),
                         );
@@ -188,9 +178,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
                         } else {
                           var filteredList = <BaseContact?>[];
                           for (var c in snapshot.data!) {
-                            if (c!.contact!.atSign!
-                                .toUpperCase()
-                                .contains(searchText.toUpperCase())) {
+                            if (c!.contact!.atSign!.toUpperCase().contains(searchText.toUpperCase())) {
                               filteredList.add(c);
                             }
                           }
@@ -208,9 +196,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
                             physics: const AlwaysScrollableScrollPhysics(),
                             itemBuilder: (context, alphabetIndex) {
                               var contactsForAlphabet = <AtContact?>[];
-                              var currentChar =
-                                  String.fromCharCode(alphabetIndex + 65)
-                                      .toUpperCase();
+                              var currentChar = String.fromCharCode(alphabetIndex + 65).toUpperCase();
                               if (alphabetIndex == 26) {
                                 currentChar = 'Others';
                                 for (var c in filteredList) {
@@ -222,8 +208,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
                                 }
                               } else {
                                 for (var c in filteredList) {
-                                  if (c!.contact!.atSign![1].toUpperCase() ==
-                                      currentChar) {
+                                  if (c!.contact!.atSign![1].toUpperCase() == currentChar) {
                                     contactsForAlphabet.add(c.contact!);
                                   }
                                 }
@@ -247,8 +232,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
                                       SizedBox(width: 4.toWidth),
                                       Expanded(
                                         child: Divider(
-                                          color: ColorConstants.dividerColor
-                                              .withOpacity(0.2),
+                                          color: ColorConstants.dividerColor.withValues(alpha: 0.2),
                                           height: 1.toHeight,
                                         ),
                                       ),
@@ -275,15 +259,14 @@ class _ContactsScreenState extends State<ContactsScreen> {
         physics: const NeverScrollableScrollPhysics(),
         shrinkWrap: true,
         separatorBuilder: (context, _) => Divider(
-              color: ColorConstants.dividerColor.withOpacity(0.2),
+              color: ColorConstants.dividerColor.withValues(alpha: 0.2),
               height: 1.toHeight,
             ),
         itemBuilder: (context, index) {
           return Padding(
             padding: const EdgeInsets.all(8.0),
             child: Slidable(
-              endActionPane:
-                  ActionPane(motion: const ScrollMotion(), children: [
+              endActionPane: ActionPane(motion: const ScrollMotion(), children: [
                 SlidableAction(
                   label: TextStrings().block,
                   backgroundColor: ColorConstants.inputFieldColor,
@@ -340,8 +323,7 @@ class _ContactsScreenState extends State<ContactsScreen> {
         ),
       ),
     );
-    await _contactService.blockUnblockContact(
-        contact: contact, blockAction: true);
+    await _contactService.blockUnblockContact(contact: contact, blockAction: true);
     setState(() {
       blockingContact = false;
       Navigator.pop(context);