+++ packages/at_onboarding_flutter/lib/screen/at_onboarding_pair_screen.dart
@@ -88,9 +88,8 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
             child: SingleChildScrollView(
               child: Container(
                 decoration: BoxDecoration(
-                    color: theme.primaryColor.withOpacity(0.1),
-                    borderRadius:
-                        BorderRadius.circular(AtOnboardingDimens.borderRadius)),
+                    color: theme.primaryColor.withValues(alpha: 0.1),
+                    borderRadius: BorderRadius.circular(AtOnboardingDimens.borderRadius)),
                 padding: const EdgeInsets.all(AtOnboardingDimens.paddingNormal),
                 margin: const EdgeInsets.all(AtOnboardingDimens.paddingNormal),
                 constraints: const BoxConstraints(
@@ -114,8 +113,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
                       focusNode: _focusNode,
                       validator: (String? value) {
                         if ((value ?? '').isEmpty) {
-                          return AtOnboardingLocalizations
-                              .current.msg_atSign_cannot_empty;
+                          return AtOnboardingLocalizations.current.msg_atSign_cannot_empty;
                         }
                         return null;
                       },
@@ -123,9 +121,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
                       inputFormatters: <TextInputFormatter>[
                         LengthLimitingTextInputFormatter(80),
                         // This inputFormatter function will convert all the input to lowercase.
-                        TextInputFormatter.withFunction(
-                            (TextEditingValue oldValue,
-                                TextEditingValue newValue) {
+                        TextInputFormatter.withFunction((TextEditingValue oldValue, TextEditingValue newValue) {
                           return newValue.copyWith(
                             text: newValue.text.toLowerCase(),
                             selection: newValue.selection,
@@ -136,8 +132,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
                       decoration: InputDecoration(
                         fillColor: Colors.blueAccent,
                         errorStyle: const TextStyle(fontSize: 12),
-                        prefixStyle:
-                            TextStyle(color: theme.primaryColor, fontSize: 15),
+                        prefixStyle: TextStyle(color: theme.primaryColor, fontSize: 15),
                         border: OutlineInputBorder(
                           borderSide: BorderSide(
                             color: theme.primaryColor,
@@ -148,8 +143,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
                             color: Colors.grey[500]!,
                           ),
                         ),
-                        contentPadding: const EdgeInsets.symmetric(
-                            horizontal: AtOnboardingDimens.paddingSmall),
+                        contentPadding: const EdgeInsets.symmetric(horizontal: AtOnboardingDimens.paddingSmall),
                       ),
                     ),
                     const SizedBox(height: 10),
@@ -212,8 +206,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
       isParing = true;
       setState(() {});
       bool status = false;
-      status =
-          await registerPersona(widget.atSign, _emailController.text, context);
+      status = await registerPersona(widget.atSign, _emailController.text, context);
       isParing = false;
       setState(() {});
       if (status) {
@@ -236,14 +229,11 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
     }
   }
 
-  Future<bool> registerPersona(
-      String atsign, String email, BuildContext context,
-      {String? oldEmail}) async {
+  Future<bool> registerPersona(String atsign, String email, BuildContext context, {String? oldEmail}) async {
     dynamic data;
     bool status = false;
     // String atsign;
-    dynamic response = await _freeAtsignService.registerPerson(atsign, email,
-        oldEmail: oldEmail);
+    dynamic response = await _freeAtsignService.registerPerson(atsign, email, oldEmail: oldEmail);
     if (response.statusCode == 200) {
       data = response.body;
       data = jsonDecode(data);
@@ -281,10 +271,8 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
             text: TextSpan(
               children: <InlineSpan>[
                 TextSpan(
-                  style: const TextStyle(
-                      color: Colors.black, fontSize: 16, letterSpacing: 0.5),
-                  text:
-                      AtOnboardingLocalizations.current.msg_maximum_atSign_prev,
+                  style: const TextStyle(color: Colors.black, fontSize: 16, letterSpacing: 0.5),
+                  text: AtOnboardingLocalizations.current.msg_maximum_atSign_prev,
                 ),
                 TextSpan(
                     text: 'https://my.atsign.com',
@@ -296,14 +284,12 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
                     recognizer: TapGestureRecognizer()
                       ..onTap = () async {
                         String url = 'https://my.atsign.com';
-                        if (!widget.hideReferences &&
-                            await canLaunchUrl(Uri.parse(url))) {
+                        if (!widget.hideReferences && await canLaunchUrl(Uri.parse(url))) {
                           await launchUrl(Uri.parse(url));
                         }
                       }),
                 TextSpan(
-                  text:
-                      AtOnboardingLocalizations.current.msg_maximum_atSign_next,
+                  text: AtOnboardingLocalizations.current.msg_maximum_atSign_next,
                   style: const TextStyle(
                     color: Colors.black,
                     fontSize: 16,
@@ -342,13 +328,11 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
     if (result != null && result.secret != null) {
       if (!mounted) return;
       Navigator.pop(context);
-      widget.onGenerateSuccess
-          ?.call(atSign: result.atSign, secret: result.secret ?? '');
+      widget.onGenerateSuccess?.call(atSign: result.atSign, secret: result.secret ?? '');
     } else if (result != null) {
       dynamic data;
       //User choose a difference atsign to onboard
-      dynamic response =
-          await _freeAtsignService.loginWithAtsign(result.atSign);
+      dynamic response = await _freeAtsignService.loginWithAtsign(result.atSign);
       if (response.statusCode == 200) {
         data = response.body;
         data = jsonDecode(data);
@@ -369,8 +353,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
       if (result2 != null) {
         if (!mounted) return;
         Navigator.pop(context);
-        widget.onGenerateSuccess
-            ?.call(atSign: result2.atSign, secret: result2.secret ?? '');
+        widget.onGenerateSuccess?.call(atSign: result2.atSign, secret: result2.secret ?? '');
       } else {
         if (!mounted) return;
         Navigator.pop(context);